+++ lib/td_ie/ingests/ingest_execution.ex
@@ -3,15 +3,18 @@ defmodule TdIe.Ingests.IngestExecution do
   use Ecto.Schema
   import Ecto.Changeset
 
+  alias TdIe.Ingests.Ingest
+
   schema "ingest_executions" do
     field(:end_timestamp, :naive_datetime)
     field(:start_timestamp, :naive_datetime)
     field(:status, :string)
     field(:file_name, :string)
     field(:file_size, :integer)
-    field(:ingest_id, :id)
     field(:description, :string)
     field(:records, :integer)
+
+    belongs_to(:ingest, Ingest, on_replace: :update)
 
     timestamps(type: :utc_datetime)
   end