+++ lib/td_ie/ingests/ingests.ex
@@ -376,6 +376,7 @@ defmodule TdIe.Ingests do
     |> join(:left, [v], _ in assoc(v, :ingest))
     |> preload([_, i], ingest: i)
     |> order_by(asc: :version)
+    |> preload(ingest: :executions)
     |> Repo.all()
   end
 
@@ -786,6 +787,14 @@ defmodule TdIe.Ingests do
     %IngestExecution{}
     |> IngestExecution.changeset(attrs)
     |> Repo.insert()
+    |> case do
+      {:ok, ingest_execution} ->
+        refresh_ingest(ingest_execution)
+        {:ok, ingest_execution}
+
+      err ->
+        err
+    end
   end
 
   @doc """
@@ -804,6 +813,14 @@ defmodule TdIe.Ingests do
     ingest_execution
     |> IngestExecution.changeset(attrs)
     |> Repo.update()
+    |> case do
+      {:ok, ingest_execution} ->
+        refresh_ingest(ingest_execution)
+        {:ok, ingest_execution}
+
+      err ->
+        err
+    end
   end
 
   @doc """
@@ -819,7 +836,16 @@ defmodule TdIe.Ingests do
 
   """
   def delete_ingest_execution(%IngestExecution{} = ingest_execution) do
-    Repo.delete(ingest_execution)
+    ingest_execution
+    |> Repo.delete()
+    |> case do
+      {:ok, %IngestExecution{}} ->
+        refresh_ingest(ingest_execution)
+        {:ok, %IngestExecution{}}
+
+      err ->
+        err
+    end
   end
 
   @doc """
@@ -833,5 +859,41 @@ defmodule TdIe.Ingests do
   """
   def change_ingest_execution(%IngestExecution{} = ingest_execution) do
     IngestExecution.changeset(ingest_execution, %{})
+  end
+
+  def get_last_execution([]), do: %{}
+
+  def get_last_execution(executions) do
+    start_execution = last_execution(executions, :start_timestamp)
+    end_execution = last_execution(executions, :end_timestamp)
+
+    [end_execution, start_execution]
+    |> Enum.filter(& &1)
+    |> Enum.sort_by(&elem(&1, 0), {:desc, NaiveDateTime})
+    |> case do
+      [] ->
+        Map.new()
+
+      [head | _] ->
+        execution = elem(head, 0)
+        status = head |> elem(1) |> Map.get(:status)
+        %{execution: execution, status: status}
+    end
+  end
+
+  defp refresh_ingest(%IngestExecution{ingest_id: ingest_id}) do
+    unless is_nil(ingest_id) do
+      IngestLoader.refresh(ingest_id)
+    end
+  end
+
+  defp last_execution(executions, key) do
+    executions
+    |> Enum.filter(&Map.get(&1, key))
+    |> Enum.sort_by(&Map.get(&1, key), {:desc, NaiveDateTime})
+    |> case do
+      [] -> nil
+      [head | _] -> {Map.get(head, key), head}
+    end
   end
 end